+++ .gitignore
@@ -15,3 +15,4 @@ verdaccio-logs.txt
 cdk.out/
 e2e-tests
 concurrent_workspace_script_cache.json
+api-validation-temp-dir/
